requirements.txt
boto3==1.14.56
botocore==1.17.56
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
decorator==4.4.2
docutils==0.15.2
idna==2.10
jmespath==0.10.0
lxml==4.5.2
pycparser==2.20
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-telegram-bot==12.8
pyvmomi==7.0
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
tornado==6.0.4
urllib3==1.25.10
